test: reorganize test classes
