+++ gravitee-apim-repository/gravitee-apim-repository-elasticsearch/src/test/java/io/gravitee/repository/elasticsearch/configuration/YamlPropertySourceFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.repository.elasticsearch.spring;
+package io.gravitee.repository.elasticsearch.configuration;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
